# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "FlashNotes", "version": "0.1.0" },
  "paths": {
...
    "/api/v1/collections/": {
      "get": {
        "tags": ["flashcards"],
        "summary": "Read Collections",
        "operationId": "flashcards-read_collections",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CollectionList" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["flashcards"],
        "summary": "Create Collection",
        "operationId": "flashcards-create_collection",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CollectionCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Collection" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}": {
      "get": {
        "tags": ["flashcards"],
        "summary": "Read Collection",
        "operationId": "flashcards-read_collection",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Collection" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["flashcards"],
        "summary": "Update Collection",
        "operationId": "flashcards-update_collection",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CollectionUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/Collection" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["flashcards"],
        "summary": "Delete Collection",
        "operationId": "flashcards-delete_collection",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}/cards/": {
      "get": {
        "tags": ["flashcards"],
        "summary": "Read Cards",
        "operationId": "flashcards-read_cards",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/CardList" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["flashcards"],
        "summary": "Create Card",
        "operationId": "flashcards-create_card",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CardCreate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Card" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/collections/{collection_id}/cards/{card_id}": {
      "get": {
        "tags": ["flashcards"],
        "summary": "Read Card",
        "operationId": "flashcards-read_card",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Card Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Card" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["flashcards"],
        "summary": "Update Card",
        "operationId": "flashcards-update_card",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Card Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CardUpdate" } }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Card" } } }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["flashcards"],
        "summary": "Delete Card",
        "operationId": "flashcards-delete_card",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Collection Id" }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Card Id" }
          }
        ],
        "responses": {
          "204": { "description": "Successful Response" },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sessions": {
      "post": {
        "tags": ["flashcards"],
        "summary": "Start Practice Session",
        "description": "Start a new practice session for a collection",
        "operationId": "flashcards-start_practice_session",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PracticeSessionCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PracticeSession" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["flashcards"],
        "summary": "List Practice Sessions",
        "description": "List all practice sessions for the current user",
        "operationId": "flashcards-list_practice_sessions",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 0, "title": "Skip" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PracticeSessionList" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sessions/{practice_session_id}": {
      "get": {
        "tags": ["flashcards"],
        "summary": "Get Practice Session Status",
        "description": "Get practice session status and statistics",
        "operationId": "flashcards-get_practice_session_status",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "practice_session_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Practice Session Id" }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/PracticeSession" } }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sessions/{practice_session_id}/cards": {
      "get": {
        "tags": ["flashcards"],
        "summary": "List Practice Cards",
        "description": "List practice cards for a session, optionally filtering and ordering.",
        "operationId": "flashcards-list_practice_cards",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "practice_session_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Practice Session Id" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "enum": ["pending", "completed", "all"], "type": "string" },
                { "type": "null" }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "default": 100, "title": "Limit" }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "enum": ["asc", "desc", "random"], "type": "string" },
                { "type": "null" }
              ],
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PracticeCardListResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/practice-sessions/{practice_session_id}/cards/{card_id}": {
      "patch": {
        "tags": ["flashcards"],
        "summary": "Update Practice Card Result",
        "description": "Update the result (is_correct) for a practiced card.",
        "operationId": "flashcards-update_practice_card_result",
        "security": [{ "OAuth2PasswordBearer": [] }],
        "parameters": [
          {
            "name": "practice_session_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Practice Session Id" }
          },
          {
            "name": "card_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid", "title": "Card Id" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PracticeCardResultPatch" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PracticeCardResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Card": {
        "properties": {
          "front": { "type": "string", "title": "Front" },
          "back": { "type": "string", "title": "Back" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "collection_id": { "type": "string", "format": "uuid", "title": "Collection Id" }
        },
        "type": "object",
        "required": ["front", "back", "id", "collection_id"],
        "title": "Card"
      },
...
      "CardCreate": {
        "properties": {
          "front": { "type": "string", "title": "Front" },
          "back": { "type": "string", "title": "Back" },
          "prompt": {
            "anyOf": [{ "type": "string", "maxLength": 100 }, { "type": "null" }],
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": ["front", "back"],
        "title": "CardCreate"
      },
      "CardList": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/Card" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "CardList"
      },
      "CardUpdate": {
        "properties": {
          "front": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Front" },
          "back": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Back" }
        },
        "type": "object",
        "title": "CardUpdate"
      },
      "Collection": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "cards": {
            "items": { "$ref": "#/components/schemas/Card" },
            "type": "array",
            "title": "Cards"
          }
        },
        "type": "object",
        "required": ["name", "id", "user_id", "cards"],
        "title": "Collection"
      },
...
      "CollectionCreate": {
        "properties": {
          "name": { "type": "string", "title": "Name" },
          "prompt": {
            "anyOf": [{ "type": "string", "maxLength": 100 }, { "type": "null" }],
            "title": "Prompt"
          }
        },
        "type": "object",
        "required": ["name"],
        "title": "CollectionCreate"
      },
      "CollectionList": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/Collection" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "CollectionList"
      },
...
      "CollectionUpdate": {
        "properties": {
          "name": { "anyOf": [{ "type": "string" }, { "type": "null" }], "title": "Name" }
        },
        "type": "object",
        "title": "CollectionUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PracticeCardListResponse": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/PracticeCardResponse" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "PracticeCardListResponse"
      },
...
      "PracticeCardResultPatch": {
        "properties": { "is_correct": { "type": "boolean", "title": "Is Correct" } },
        "type": "object",
        "required": ["is_correct"],
        "title": "PracticeCardResultPatch"
      },
      "PracticeSession": {
        "properties": {
          "collection_id": { "type": "string", "format": "uuid", "title": "Collection Id" },
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "user_id": { "type": "string", "format": "uuid", "title": "User Id" },
          "is_completed": { "type": "boolean", "title": "Is Completed" },
          "total_cards": { "type": "integer", "title": "Total Cards" },
          "cards_practiced": { "type": "integer", "title": "Cards Practiced" },
          "correct_answers": { "type": "integer", "title": "Correct Answers" },
          "created_at": { "type": "string", "format": "date-time", "title": "Created At" },
          "updated_at": { "type": "string", "format": "date-time", "title": "Updated At" },
          "practice_cards": {
            "items": { "$ref": "#/components/schemas/PracticeCard" },
            "type": "array",
            "title": "Practice Cards"
          }
        },
        "type": "object",
        "required": [
          "collection_id",
          "id",
          "user_id",
          "is_completed",
          "total_cards",
          "cards_practiced",
          "correct_answers",
          "created_at",
          "updated_at",
          "practice_cards"
        ],
        "title": "PracticeSession"
      },
      "PracticeSessionCreate": {
        "properties": {
          "collection_id": { "type": "string", "format": "uuid", "title": "Collection Id" }
        },
        "type": "object",
        "required": ["collection_id"],
        "title": "PracticeSessionCreate"
      },
      "PracticeSessionList": {
        "properties": {
          "data": {
            "items": { "$ref": "#/components/schemas/PracticeSession" },
            "type": "array",
            "title": "Data"
          },
          "count": { "type": "integer", "title": "Count" }
        },
        "type": "object",
        "required": ["data", "count"],
        "title": "PracticeSessionList"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": { "password": { "scopes": {}, "tokenUrl": "/api/v1/tokens" } }
      }
    }
  }
}